+++ resources/views/auth/forgot-password.blade.php
@@ -9,7 +9,7 @@
         <form method="POST" action="{{ route('password.email') }}">
             @csrf
             <x-input-email for="email" label="E-mail Address" autocomplete="email" />
-            <x-submit text="EMAIL PASSWORD RESET LINK" rowClasses="tar" />
+            <x-submit text="EMAIL PASSWORD RESET LINK" rowClass="tar" />
         </form>