+++ resources/views/auth/reset-password.blade.php
@@ -5,10 +5,10 @@
         <form method="POST" action="{{ route('password.update') }}">
             @csrf
             <input type="hidden" name="token" value="{{ $request->route('token') }}">
-            <x-input-email for="email" label="E-mail Address" autocomplete="email" rowClasses="mxy-0"/>
+            <x-input-email for="email" label="E-mail Address" autocomplete="email" rowClass="mxy-0"/>
             <x-input-password for="password" label="Password" autocomplete="password" />
             <x-input-password for="password_confirmation" label="Confirm Password" autocomplete="new-password" />
-            <x-submit text="Reset Password" rowClasses="tar" />
+            <x-submit text="Reset Password" rowClass="tar" />
         </form>
 
     </x-authit::auth-box>