+++ src/InstallCommand.php
@@ -11,7 +11,7 @@ class InstallCommand extends Command
     /**
      * The name and signature of the console command.
      */
-    protected $signature = 'authit:install';
+    protected $signature = 'authit:install {--L|local : Indicates if components and views support should be published}';
 
     /**
      * The console command description.
@@ -23,6 +23,14 @@ class InstallCommand extends Command
      */
     public function handle()
     {
+
+        // publish local assets...
+        if ($this->option('local')) {
+            $this->callSilent('vendor:publish', ['--tag' => 'authit-views', '--force' => true]);
+        }
+
+        // Publish spatie permissions and update kernel.php...
+        $this->installPermissions();
 
         // Nav...
         (new Filesystem)->ensureDirectoryExists(resource_path('navs'));
@@ -53,6 +61,36 @@ class InstallCommand extends Command
         }
 
         return Command::SUCCESS;
+    }
+
+
+    public function installPermissions()
+    {
+
+        $this->callSilent('vendor:publish', ['--provider' => 'Spatie\Permission\PermissionServiceProvider', '--force' => true]);
+
+        // // Include HasRoles trait in user model
+        if (!$this->stringInFile('./app/Models/User.php', "HasRoles")) {
+            $this->replaceInFile('HasFactory,', 'HasFactory, HasRoles,', 'app/Models/User.php');
+
+            $this->replaceInFile(
+                'use Illuminate\Database\Eloquent\Factories\HasFactory;',
+                "use Illuminate\Database\Eloquent\Factories\HasFactory;\ruse Spatie\Permission\Traits\HasRoles;",
+                'app/Models/User.php'
+            );
+        }
+
+        // Add middleware to kernel.php
+        if (!$this->stringInFile('./app/Http/kernel.php', "Spatie\Permission\Middlewares")) {
+            $this->replaceInFile(
+                "'verified' => \Illuminate\Auth\Middleware\EnsureEmailIsVerified::class,",
+                "'verified' => \Illuminate\Auth\Middleware\EnsureEmailIsVerified::class,
+        'role' => \Spatie\Permission\Middlewares\RoleMiddleware::class,
+        'permission' => \Spatie\Permission\Middlewares\PermissionMiddleware::class,
+        'role_or_permission' => \Spatie\Permission\Middlewares\RoleOrPermissionMiddleware::class,",
+                './app/Http/kernel.php'
+            );
+        }
     }
 
     /**