+++ src/routes.php
@@ -55,3 +55,4 @@ Route::middleware(['web', 'auth'])->group(function () {
 
     Route::post('logout', [AuthenticatedSessionController::class, 'destroy'])->name('logout');
 });
+